Moved time_consuming common functions to own file
