+++ tests/time_consuming_test.rs
@@ -1,10 +1,10 @@
-use std::{collections::HashMap, path::Path};
+use std::path::Path;
 mod ef_tests_executor;
 use ef_tests_executor::{
-    models::{AccountInfo, TestSuite},
-    test_utils::decode_hex,
+    models::TestSuite,
+    test_utils::{setup_evm, verify_result, verify_storage},
 };
-use evm_mlir::{db::Db, env::TransactTo, result::ExecutionResult, Env, Evm};
+use evm_mlir::env::TransactTo;
 
 fn run_test(path: &Path, contents: String) -> datatest_stable::Result<()> {
     let test_suite: TestSuite = serde_json::from_reader(contents.as_bytes())
@@ -19,96 +19,15 @@ fn run_test(path: &Path, contents: String) -> datatest_stable::Result<()> {
             Some(to) => TransactTo::Call(to),
             None => TransactTo::Create,
         };
-
         let sender = unit.transaction.sender.unwrap_or_default();
         let gas_price = unit.transaction.gas_price.unwrap_or_default();
 
         for test in tests {
-            let mut env = Env::default();
-            env.tx.transact_to = to.clone();
-            env.tx.gas_price = gas_price;
-            env.tx.caller = sender;
-            env.tx.gas_limit = unit.transaction.gas_limit[test.indexes.gas].as_u64();
-            env.tx.value = unit.transaction.value[test.indexes.value];
-            env.tx.data = decode_hex(unit.transaction.data[test.indexes.data].clone()).unwrap();
-
-            env.block.number = unit.env.current_number;
-            env.block.coinbase = unit.env.current_coinbase;
-            env.block.timestamp = unit.env.current_timestamp;
-            let excess_blob_gas = unit
-                .env
-                .current_excess_blob_gas
-                .unwrap_or_default()
-                .as_u64();
-            env.block.set_blob_base_fee(excess_blob_gas);
-
-            if let Some(basefee) = unit.env.current_base_fee {
-                env.block.basefee = basefee;
-            };
-            let mut db = match to.clone() {
-                TransactTo::Call(to) => {
-                    let opcodes = decode_hex(unit.pre.get(&to).unwrap().code.clone()).unwrap();
-                    Db::new().with_contract(to, opcodes)
-                }
-                TransactTo::Create => {
-                    let opcodes =
-                        decode_hex(unit.pre.get(&env.tx.caller).unwrap().code.clone()).unwrap();
-                    Db::new().with_contract(env.tx.get_address(), opcodes)
-                }
-            };
-
-            // Load pre storage into db
-            for (address, account_info) in unit.pre.iter() {
-                let opcodes = decode_hex(account_info.code.clone()).unwrap();
-                db = db.with_contract(address.to_owned(), opcodes);
-                db.set_account(
-                    address.to_owned(),
-                    account_info.nonce,
-                    account_info.balance,
-                    account_info.storage.clone(),
-                );
-            }
-            let mut evm = Evm::new(env, db);
-
+            let mut evm = setup_evm(test, &unit, &to, sender, gas_price);
             let res = evm.transact().unwrap();
-
-            match (&test.expect_exception, &res.result) {
-                (None, _) => {
-                    if unit.out.as_ref() != res.result.output() {
-                        return Err("Wrong output".into());
-                    }
-                }
-                (Some(_), ExecutionResult::Halt { .. } | ExecutionResult::Revert { .. }) => {
-                    return Ok(()); //Halt/Revert and want an error
-                }
-                _ => {
-                    return Err("Expected exception but got none".into());
-                }
-            }
-
+            verify_result(test, unit.out.as_ref(), &res.result)?;
             // TODO: use rlp and hash to check logs
-
-            // Test the resulting storage is the same as the expected storage
-            let mut result_state = HashMap::new();
-            for address in test.post_state.keys() {
-                let account = res.state.get(address).unwrap();
-                let opcodes = decode_hex(account.info.code.clone().unwrap()).unwrap();
-                result_state.insert(
-                    address.to_owned(),
-                    AccountInfo {
-                        balance: account.info.balance,
-                        code: opcodes,
-                        nonce: account.info.nonce,
-                        storage: account
-                            .storage
-                            .clone()
-                            .into_iter()
-                            .map(|(addr, slot)| (addr, slot.present_value))
-                            .collect(),
-                    },
-                );
-            }
-            assert_eq!(test.post_state, result_state);
+            verify_storage(&test.post_state, res.state);
         }
     }
     Ok(())
